requirements.txt
Flask==2.3.3
Pillow==10.0.1
transformers>=4.55.0
torch>=2.7.0
torchaudio>=2.7.0
torchvision>=0.22.0
diffusers>=0.36.0
peft>=0.17.0
accelerate>=1.10.0
safetensors>=0.6.0
huggingface_hub>=0.34.0